pyproject.toml
[project]
name = "nevup-ai-engine"
version = "0.1.0"
description = "NevUp Hackathon 2026 — Track 2: System of AI Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pyjwt>=2.8.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "python-json-logger>=2.0.0",
    "sse-starlette>=2.0.0",
    "huggingface-hub>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
